refactor: Move badge & streamder data request
